feat: add `CreatePrismaOptionsFactory` update`ConfigurableModuleBuilder`
This commit introduces the `CreatePrismaOptionsFactory` interface, which
defines a contract for classes responsible for creating Prisma module
options. This improves code organization, testability, and promotes
dependency injection.

The commit also updates the `ConfigurableModuleBuilder` to utilize this
interface. Specifically, the `.setFactoryMethodName('createPrismaOptions')`
is added to properly configure the builder to use the
`createPrismaOptions` method of classes implementing the
`CreatePrismaOptionsFactory` interface.